js/xpconnect/tests/chrome/test_bug760076.xul

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=760076
-->
<window title="Mozilla Bug 760076"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=760076"
     target="_blank">Mozilla Bug 760076</a>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript">
  <![CDATA[
  /** Test for constructing COW-ed functions in content. **/

  // This gets decompiled and run in the sandbox.
  function sandboxCode() {
    try {
      is(chromeFunction(), 42, "Should call successfully");
    } catch(e) { ok(false, "Shouldn't throw when calling"); }

    try {
      ok(typeof (new chromeFunction()) !== 'undefined',
         "Should construct successfully");
    } catch(e) { ok(false, "Shouldn't throw when constructing"); }
  }

  // Set up the sandbox.
  const Cu = Components.utils;
  var sb = new Cu.Sandbox('http://www.example.com');

  // Import the functions it needs.
  sb.ok = ok;
  sb.is = is;
  sb.chromeFunction = function() { ok(true, "Called chrome function"); return 42; }
  Cu.evalInSandbox(sandboxCode.toSource(), sb);

  // Do the test.
  Cu.evalInSandbox("sandboxCode();", sb);

  ]]>
  </script>
</window>
